==> notebooks/NB1_data_gathering.ipynb <==
Overview

In this notebook, we have a script that scrapes various data pertaining to import and export information for each country from 1995 to 2012, as well as stock index data for a selection of developed nations.  For each country whose import and export data we scrape, we are collecting data on who they import and export from, and the mix of products that they import and export.  For each country whose stock index data we scrape, we pull raw prices of the country's main stock index for the years under investigation.  After scraping the data in the form of Python dictionaries, we conduct some basic scrubbing techniques before storing the relevant information in CSVs for later use.  Furthermore, to demonstrate analysis techniques using numpy, we create the Trade Diversification Index before storing that as our measure of trade diversity. We will be going over the data gathering process of our 4 datasets:
<br />
<br />
1. Export and Import Data
<br />
2. Currency Exchange Rates
<br />
3. Stock Index Prices
<br />
4. Economic Indicators

Before we proceed, for simplicity, we will import all the libraries that we will need in this notebook.

import json
import csv
import string
import requests
import os
from pandas import DataFrame
from numpy import array, ndarray, matrix
import numpy
import ystockquote

Export and Import Data

Export and Import data is collected through scraping the Obervatory of Economic Complexity website from MIT's Media Lab. We are gathering data on the monetary values of goods being exported and imported, as well as to which partner countries.

The following function definitions in the Python language are aimed to do just that. We first define the set of countries we wish to gather data from.

def load_country_dict():
    """
    Here we are loading from a CSV the countries about which we are able to scrape data.  This list of countries was downloaded directly from http://atlas.media.mit.edu/about/data/country/.  The goal of this project was to scrape the import/export data, so we simply downloaded this data from the website.
    """
    f = open('country_codes.txt', 'rU')
    reader = csv.reader(f)
    country_dict = {}
    for row in reader:
        country_dict[row[0]] = row[1]
    return country_dict

Now, we are ready to define the functions that will perform the actual scraping. For your reference, the website is http://atlas.media.mit.edu/hs/export/

def get_product_data(year, country):
    """
    This function returns the JSON file that contains the data on each country from the URL specified below.  The JSON file is pulled as a dictionary, and is then scrubbed into a list of dictionaries, in which each dictionary contains information about a particular product that is imported or exported.
    """
    
    url = 'http://atlas.media.mit.edu/hs/export/' + str(year) + '/' + country + '/all/show/'
    page = requests.get(url)
    return page.json().values()[0]

def write_data_to_files():
    """
    This function uses get_product_data() and get_partner_data() to write files containing the data scraped from the Observatory of Economic Complexity.  These will be formatted as CSVs for easy access later.
    """
    
    country_dict = load_country_dict()
    countries = country_dict.keys()
    years = range(1995, 2012)
    for country in countries:
        os.makedirs('data/part/' + country)
        os.makedirs('data/prod/' + country)
        for year in years:
            f = open('data/part/' + country + '/' + str(year) + '.txt', 'wb')
            writer = csv.writer(f)
            to_write = []
            part_raw_data = get_partner_data(year, country)
            for i in range(len(part_raw_data)):
                temp = []
                temp.append(part_raw_data[i]['dest_id'])
                if 'export_val' in part_raw_data[i].keys():
                    temp.append(part_raw_data[i]['export_val'])
                else:
                    temp.append(0)
                if 'import_val' in part_raw_data[i].keys():
                    temp.append(part_raw_data[i]['import_val'])
                else:
                    temp.append(0)
                to_write.append(temp)
            writer.writerows(to_write)
            f.close()
            f = open('data/prod/' + country + '/' + str(year) + '.txt', 'wb')
            writer = csv.writer(f)
            to_write = []
            prod_raw_data = get_product_data(year, country)
            for i in range(len(prod_raw_data)):
                temp = []
                temp.append(prod_raw_data[i]['hs_id'])
                if 'export_val' in prod_raw_data[i].keys():
                    temp.append(prod_raw_data[i]['export_val'])
                else:
                    temp.append(0)
                if 'import_val' in prod_raw_data[i].keys():
                    temp.append(prod_raw_data[i]['import_val'])
                else:
                    temp.append(0)
                to_write.append(temp)
            writer.writerows(to_write)
            f.close()
            
def create_export_product_csv():
    years = ['1995', '1996', '1997', '1998', '1999', '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008', '2009', '2010', '2011']
    countries = load_country_dict().keys()
    values_arr = []
    countries_proxy = []
    for country in countries:
        countries_proxy.append(country)
    values_arr.append(countries_proxy)
    for year in years:
        year_arr = []
        for country in countries:
            year_arr.append(calc_export_product_variance(year, country))
        values_arr.append(year_arr)
    f = open('data/trade/export_prod.csv', 'wb')
    writer = csv.writer(f)
    writer.writerows(values_arr)
    f.close()
    
def create_import_product_csv():
    years = ['1995', '1996', '1997', '1998', '1999', '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008', '2009', '2010', '2011']
    countries = load_country_dict().keys()
    values_arr = []
    countries_proxy = []
    for country in countries:
        countries_proxy.append(country)
    values_arr.append(countries_proxy)
    for year in years:
        year_arr = []
        for country in countries:
            year_arr.append(calc_import_product_variance(year, country))
        values_arr.append(year_arr)
    f = open('data/trade/import_prod.csv', 'wb')
    writer = csv.writer(f)
    writer.writerows(values_arr)
    f.close()
    
def create_export_partner_csv():
    years = ['1995', '1996', '1997', '1998', '1999', '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008', '2009', '2010', '2011']
    countries = load_country_dict().keys()
    values_arr = []
    countries_proxy = []
    for country in countries:
        countries_proxy.append(country)
    values_arr.append(countries_proxy)
    for year in years:
        year_arr = []
        for country in countries:
            year_arr.append(calc_export_partner_variance(year, country))
        values_arr.append(year_arr)
    f = open('data/trade/export_part.csv', 'wb')
    writer = csv.writer(f)
    writer.writerows(values_arr)
    f.close()
    
def create_import_partner_csv():
    years = ['1995', '1996', '1997', '1998', '1999', '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008', '2009', '2010', '2011']
    countries = load_country_dict().keys()
    values_arr = []
    countries_proxy = []
    for country in countries:
        countries_proxy.append(country)
    values_arr.append(countries_proxy)
    for year in years:
        year_arr = []
        for country in countries:
            year_arr.append(calc_import_partner_variance(year, country))
        values_arr.append(year_arr)
    f = open('data/trade/import_part.csv', 'wb')
    writer = csv.writer(f)
    writer.writerows(values_arr)
    f.close()
    

The above functions neatly organized the scraped export and import data into .csv files, which will be easy to load into more complex functions that we will define now:
<br />
<br />
1. **load_country_product_data**: this function simply loads our .csv files and returns an array of the products imported and exported by a given country for a given year.

def load_country_product_data(year, country):
    """
    This function explores the data that has been previously scraped and written to files by the function write_data_to_files(), and returns an array of the products imported and exported by a given country for a given year.
    """
    
    f = open('data/prod/' + country + '/' + str(year) + '.txt', 'rU')
    reader = csv.reader(f)
    product_array = []
    for row in reader:
        product_array.append(row)
    f.close()
    return product_array

'2. **load_country_partner_data**: this function explores the data that has been previously scraped and written to files by the function write_data_to_files(), and loads an array of the import and export partners for a given country for a given year.

def load_country_partner_data(year, country):
    """
    This function explores the data that has been previously scraped and written to files by the function write_data_to_files(), and loads an array of the import and export partners for a given country for a given year.
    """
    
    f = open('data/part/' + country + '/' + str(year) + '.txt', 'rU')
    reader = csv.reader(f)
    partner_array = []
    for row in reader:
        partner_array.append(row)
    f.close()
    return partner_array

'3. **get_product_breakdown**: This function breaks down each product to give us the percentage mix of a given country's imports and exports for a given year.

def get_product_breakdown(year, country):
    """
    This function breaks down each product to give us the percentage mix of a given country's imports and exports for a given year.
    """
    
    product_array = load_country_product_data(year, country)
    total_exports = 0
    total_imports = 0
    for i in range(len(product_array)):
        total_exports += float(product_array[i][1])
        total_imports += float(product_array[i][2])
    product_mix_array = []
    for i in range(len(product_array)):
        if total_exports != 0 and total_imports != 0:
            temp = []
            temp.append(product_array[i][0])
            temp.append(float(product_array[i][1])/float(total_exports)*float(100))
            temp.append(float(product_array[i][2])/float(total_imports)*float(100))
            product_mix_array.append(temp)
        else:
            print country
    return product_mix_array


'4. **get_partner_breakdown**: This function breaks down each partner to give us the percentage mix of a given country's imports and exports for a given year.

def get_partner_breakdown(year, country):
    """
    This function breaks down each partner to give us the percentage mix of a given country's imports and exports for a given year.
    """
    
    partner_array = load_country_partner_data(year, country)
    total_exports = 0
    total_imports = 0
    partner_return = []
    for i in range(len(partner_array)):
        total_exports += float(partner_array[i][1])
        total_imports += float(partner_array[i][2])
    product_mix_array = []
    for i in range(len(partner_array)):
        if total_exports != 0 and total_imports != 0:
            temp = []
            temp.append(partner_array[i][0])
            temp.append(float(partner_array[i][1])/float(total_exports)*float(100))
            temp.append(float(partner_array[i][2])/float(total_imports)*float(100))
            partner_return.append(temp)
        else:
            print country
    return partner_return

'5. **categorize_products**: This function breaks down the product mix into the 21 categories provided in order to find the diversity of trade. This is necessary because the types of goods in our dataset is simply way too many to analyze in a clear manner. These 21 categories are used as standards for categorizing traded goods around the world.

def categorize_products(year, country):
    """
    This function breaks down the product mix into the 21 categories provided in order to find the diversity of trade.
    """
    
    product_mix_array = get_product_breakdown(year, country)
    product_export_mix_dict = {}
    product_import_mix_dict = {}
    for i in range(len(product_mix_array)):
        hs4_id = product_mix_array[i][0]
        if hs4_id[0:2] in product_export_mix_dict.keys():
            product_export_mix_dict[hs4_id[0:2]] += float(product_mix_array[i][1])
        else:
            product_export_mix_dict[hs4_id[0:2]] = float(product_mix_array[i][1])
        if hs4_id[0:2] in product_import_mix_dict.keys():
            product_import_mix_dict[hs4_id[0:2]] += float(product_mix_array[i][2])
        else:
            product_import_mix_dict[hs4_id[0:2]] = float(product_mix_array[i][2])
    return product_export_mix_dict, product_import_mix_dict

'6. **categorize_partners**: This function breaks down the partner mix into the various continents provided in order to find the diversities of import partners and export partners.  This is necessary because the number of countries in our dataset is to high to analyze in a clear manner.  We have defined these continent "categories" as our way of categorizing import and export partners.

def categorize_partners(year, country):
    """
    This function breaks down the partner mix into the 6 main continents provided in order to find the diversity of trade.
    """
    
    partner_mix_array = get_partner_breakdown(year, country)
    partner_export_mix_dict = {}
    partner_import_mix_dict = {}
    for i in range(len(partner_mix_array)):
        dest_id = partner_mix_array[i][0]
        if dest_id[0:2] in partner_export_mix_dict.keys():
            partner_export_mix_dict[dest_id[0:2]] += float(partner_mix_array[i][1])
        else:
            partner_export_mix_dict[dest_id[0:2]] = float(partner_mix_array[i][1])
        if dest_id[0:2] in partner_import_mix_dict.keys():
            partner_import_mix_dict[dest_id[0:2]] += float(partner_mix_array[i][2])
        else:
            partner_import_mix_dict[dest_id[0:2]] = float(partner_mix_array[i][2])
    return partner_export_mix_dict, partner_import_mix_dict

Trade Diversification Index (TDI)

We will sidetrack a little to cover the next couple of functions. As mentioned, our project is to analyze the effects of the diversification of a given country's imported and exported goods/partners. Basically, we have 4 sub-datasets within export and import data -- amount of imported goods (for each category), amount of exported goods (for each category), amount of imported goods (for each partner country), amount of exported goods (for each partner country).
<br />
<br />
With these, and associating diversification with variance, we will determine the variance of each sub-dataset, and then sum them up by using equal weights. The resulting value will be a given country's TDI.

def calc_export_product_variance(year, country):
    breakdown = categorize_products(year, country)[0].values()
    for i in range(len(breakdown)):
        breakdown[i] = round(breakdown[i], 2)
    breakdown = array(breakdown)
    return numpy.var(breakdown)

def calc_import_product_variance(year, country):
    breakdown = categorize_products(year, country)[1].values()
    for i in range(len(breakdown)):
        breakdown[i] = round(breakdown[i], 2)
    breakdown = array(breakdown)
    return numpy.var(breakdown)

def calc_export_partner_variance(year, country):
    breakdown = categorize_partners(year, country)[0].values()
    for i in range(len(breakdown)):
        breakdown[i] = round(breakdown[i], 2)
    breakdown = array(breakdown)
    return numpy.var(breakdown)

def calc_import_partner_variance(year, country):
    breakdown = categorize_partners(year, country)[1].values()
    for i in range(len(breakdown)):
        breakdown[i] = round(breakdown[i], 2)
    breakdown = array(breakdown)
    return numpy.var(breakdown)

def calc_trade_index(year, country):
    export_product_variance = calc_export_product_variance(year, country)
    import_product_variance = calc_import_product_variance(year, country)
    export_partner_variance = calc_export_partner_variance(year, country)
    import_partner_variance = calc_import_partner_variance(year, country)
    variances = array([export_product_variance, import_product_variance, export_partner_variance, import_partner_variance])
    index = numpy.mean(variances)
    return index


Currency Exchange Rates

For currency exchange, we are fortunate to find a site that can give us historical exchange rates for currencies across the world in .csv files. The website url is http://www.oanda.com/currency/historical-rates/ . We simply refer to the countries that already have import and export data above, and download the .csv file from the site if available.

Stock Index Prices

Again, from the countries that already have import and export data, we try to obtain historical prices for those countries' major stock composite indexes. The difficulty we face is that many smaller, less well-known countries do not have such information publicly (or freely) available. Furthermore, even though the Dow Jones have indexes for quite a number of the more major countries, they have commercialized their data and we are unable to obtain them at no cost. Nevertheless, we are still able to obtain the historical prices (from 1995-2011, subject to availability) for 25 major countries. The pool of countries is also diversified, and so we feel that this will still be sufficient for analysis.
<br />
<br />
We essentially have 2 sources of obtaining historical stock index prices. The first is from respective stock exchange websites where we are able to download the .csv files manually. The other is from inquiring Yahoo! Finance from within Python itself (if the ticker symbol is present on the site).

Manual .csv Files

After having downloaded the .csv files manually, we now load them into Python and extract only the essential information that we need. For instance, the .csv files give us opening price, closing price, highs and lows on a daily basis. We just need a single price for each given year in 1995-2011. We store them into .csv files again, containing just average annual prices against years from 1995-2011.

def csvExtract(filename, country):
    """
    This function takes a previously downloaded csv file of the historical daily prices for a stock index, and determines the average price for each year, storing the averages in another csv file with the designated country's name.
    """
    
    rawdata = []
    filepath = 'data/stock/manual/' + filename
    file2 = open(filepath, 'rU')
    reader2 = csv.reader(file2)
    for row in reader2:
        rawdata.append(row)

	currentYear = rawdata[1][0][-2:]
	prices = []
	data = []
	for i in range(1, len(rawdata)):
		if rawdata[i][0][-2:] == currentYear:
			prices.append(float(rawdata[i][4]))
		else:
			data.append([currentYear, float(sum(prices))/len(prices)])
			prices = []
			currentYear = rawdata[i][0][-2:]
			prices.append(float(rawdata[i][4]))

	count = 0
	for row in data:
		if count <= 12:
			row[0] = '20' + row[0]
			count += 1
		elif count >= 12:
			row[0] = '19' + row[0]
	data = sorted(data)

	with open('data/stock/done/' + country + '.csv', 'w') as file1:
		file1write = csv.writer(file1, delimiter = ',')
		file1write.writerows(data)

	file1.close()


Yahoo! Finance

For inquiring stock data from Yahoo! finance, there is a very neat Python library **ystockquote**, which returns historical prices in nested dictionaries. Again, for uniformity, we extracted the data that we need (since we are given daily opening/closing/high/low prices), and store them into .csv files with merely annual average prices against years from 1995-2011.

def yahooExtract(ticker, country):
    """
    This function uses the python library 'ystockquote', which requests historical prices and returns them as dictionaries.  We then determine the average prices for each year, storing these averages in a csv file with the given country's name.
    """
    
    rawdata = ystockquote.get_historical_prices(ticker, '1995-01-01', '2011-12-31')
    
    prices = {}
    for date in rawdata.keys():
        if date[0:4] in prices.keys():
            prices[date[0:4]].append(float(rawdata[date]['Close']))
        else:
			prices[date[0:4]] = []
			prices[date[0:4]].append(float(rawdata[date]['Close']))

	data = []
	for year in sorted(prices.keys()):
		data.append([year, float(sum(prices[year]))/len(prices[year])])

	with open('data/stock/done/' + country + '.csv', 'w') as file1:
		file1write = csv.writer(file1, delimiter = ',')
		file1write.writerows(data)

	file1.close()

We have now defined both functions that we will need to gather data from our 2 types of data sources for stock index prices. We have also predefined on a .csv file which of the 2 methods we will be using to collect the data from, as well as noting if their data is unattainable. The following code will write final .csv files into a single folder, where each .csv file is a given country's historical index price.

def getIndexPrices():
    """
    Now that we have defined all of the functions required, we run through a previously defined csv file that links each country to its stock index data source.  We also use the standard 3-letter country codes when making new csv files.
    """
    
    filename = 'data/stock/country_indexes.csv'
    
    indexes = {}
    
    csvfile = open(filename, 'rU')
    reader = csv.reader(csvfile)
    
    for row in reader:
        indexes[row[0]] = row[1:]
    countryCodes = {}
    csvfile2 = open('data/stock/country_codes.txt', 'rU')
    reader2 = csv.reader(csvfile2)
    for row in reader2:
        countryCodes[row[1].lower()] = row[0]
    
    csvfile2.close()
    
    for country in indexes.keys():
        countryCode = countryCodes[country.lower()]
        if indexes[country][1] == 'manual':
            csvExtract(indexes[country][2], countryCode)
        else:
            yahooExtract(indexes[country][1], countryCode)
    csvfile.close()

Economic Indicators

Our last dataset is one of macroeconomic indicators.  Like currencies, we are able to manually download .csv files from the World Bank's database for World Development Indicators.  We have a lot of missing data, but we will cover that in the data cleaning portion of our project.

Conclusion

To recap, since we have numerous datasets and sources, we have now gathered the 4 required datasets for our analysis (in raw form) as .csv files. We do this because we would like to use both Python and R extensively, for data gathering, data cleaning, and data analysis, and having gathered data in the form of .csv files will give a seamless transition between the 2 programming languages. This way, we can utilize the strengths of different programming languages to our own advantage (Python libraries such as ystockquote, and R's statistical analysis tools).

References

* The Observatory of Economic Complexity. 
    * Accessed with http://atlas.media.mit.edu/

* The World Bank: World Development Indicators. 
    * Accessed with http://databank.worldbank.org/data/views/variableSelection/selectvariables.aspx?source=world-development-indicators

* Oanda Currency Converter. 
    * Accessed with http://www.oanda.com/currency/converter/
