==> notebooks/NB2_data_cleaning.ipynb <==
Overview

This R Script will take the scraped data from our CSV files and put it into R Data Frames for easy analysis.  The output of the final functions will be a Data Frame with years for rows and countries for columns.  Additionally, we will make some basic plots to identify areas of analysis that we should explore more thoroughly.  Here we also introduce our macroeconomic data and foreign exchange rate data, both of which were scraped manually due to a simple interface provided by our source websites.

%load_ext rmagic

Import and Export Data

We will first gather the trade data that we have stored in the csv files. Each of the csv files contains the variance of the particular category for each country in each year.  We will gather the trade data in order to understand how diversity of trade affects countries' economic performance.
<br />
<br />
For import and export data, most of the data cleaning has been done largely during the data gathering phase, since we scrapped the website to collect the just the data that we need.

%%R

gatherTradeData = function(filename) {
  trade = read.csv(filename)
  years = 1995:2011
  rownames(trade) = years
  return(trade)
}

getExportPartnerVariances = function() {
  return(gatherTradeData("./data/trade/export_part.csv"))
}
getExportProductVariances = function() {
  return(gatherTradeData("./data/trade/export_prod.csv"))
}
getImportPartnerVariances = function() {
  return(gatherTradeData("./data/trade/import_part.csv"))
}
getImportProductVariances = function() {
  return(gatherTradeData("./data/trade/import_prod.csv"))
}
getTradeIndex = function() {
  return(gatherTradeData("./data/trade/trade_index.csv"))
}

Currency Exchange Rates

Now we will pull the exchange rate data from the csv files we have gathered from Oanda, an open source ForEx database.
<br />
<br />
This dataset is probably the most complex for analysis to be done during the next stage.  As you might have guessed, the nominal values we have gathered in our data are ratios between each country's nominal currency and the US Dollar, but because the exchange rate is a ratio rather than a raw currency value, we are able to compare the exchange rates across currencies.  We also set the entries to NA for any data that was not available.

%%R

gatherExchangeData = function() {
  exchangePath = "./data/exchange_rate/"
  exchangeFiles = list.files(exchangePath)
  
  exchangeCountries = c()
  exchangeNew = TRUE
  for (i in 1:length(exchangeFiles)) {
    exchangeCountries[i] = substring(exchangeFiles[i], 1, 3)
    exchangeFilePath = paste("./data/exchange_rate/", exchangeFiles[i], sep="")
    exchangeRawData = read.csv(exchangeFilePath, header = FALSE)
    exchangeRawData = exchangeRawData[-1,]
    exchangeRawData = exchangeRawData[-1,]
    exchangeRawData = exchangeRawData[-1,]
    exchangeRawData = exchangeRawData[-dim(exchangeRawData)[1],]
    exchangeRawData = exchangeRawData[-dim(exchangeRawData)[1],]
    exchangeRawData = exchangeRawData[-dim(exchangeRawData)[1],]
    exchangeRawData = exchangeRawData[c(dim(exchangeRawData)[1]:1),]
    exchangeRawData = exchangeRawData[2]
    if (dim(exchangeRawData)[1]<17) {
      dummyYears = 1:(17-dim(exchangeRawData)[1])
      V2 = rep(NaN, length(dummyYears))
      dummyData = data.frame(V2)
      exchangePrices = rbind(dummyData, exchangeRawData)
    }
    else {
      exchangePrices = exchangeRawData
    }
    if (exchangeNew == TRUE) {
      exchanges = exchangePrices
      exchangeNew = FALSE
    }
    else {
      exchanges = cbind(exchanges, exchangePrices)
    }
  }
  
  years = 1995:2011
  rownames(exchanges) = years
  colnames(exchanges) = exchangeCountries
  
  return(exchanges)
}

Stock Index Prices

In this case, the historical prices are given in the respective countries' own currencies.  Here we focus on **returns** year-over-year for easy analysis across countries.
<br />
<br />
To give more explanation on how we do this, we take the first year data is available (from 1995-2011) as the **base year**, and we replace it with a value of a 100, storing the real nominal value. Then, for each subsequent year, we normalize them by taking as a fraction of the base year's real nominal value. This "cache-and-step" process allows us to quickly and efficiently move from noisy nominal values to clean, normalized data that easily reveals the returns each year. Again, we set NA for data that is not available (some stock exchanges were established after 1995).

%%R

gatherIndexData = function() {
  # This function will gather the stock index data for each of the 25 countries whose
  # index data we have scraped and placed in csv files.  We will pull the data from
  # each csv and clean it such that we have relative changes rather than nominal values
  
  indexFiles = list.files(path = "./data/stock/done")
  indexCountries = c()
  new = TRUE
  for (i in 1:length(indexFiles)) {
    indexCountries[i] = substring(indexFiles[i], 1, 3)
    filePath = paste("./data/stock/done/", indexFiles[i], sep="")
    rawData = read.csv(filePath, header=FALSE)
    startYear = rawData[1][1,]
    if (startYear != 1995) {
      dummyYears = 1995:(startYear-1)
      V2 = rep(NaN, length(dummyYears))
      dummyData = data.frame(V2)
      prices = rbind(dummyData, rawData[2])
    }
    else {
      prices = rawData[2]
    }
    if (new == TRUE) {
      indexes = prices
      new = FALSE
    }
    else {
      indexes = cbind(indexes, prices)
    }
  }
  
  years = 1995:2011
  rownames(indexes) = years
  colnames(indexes) = indexCountries

for (i in colnames(indexes)) {
    k = 1995
    while (as.character(indexes[i][as.character(k),]) == "NaN") {
      k = k + 1
    }
    
    base = indexes[i][as.character(k),]
    indexes[i][as.character(k),] = NaN
    for (j in (k+1):2011) {
      temp = indexes[i][as.character(j),]
      indexes[i][as.character(j),] = indexes[i][as.character(j),]/base*100
      base = temp
    }
  }
  indexes2 = indexes[-1,]
  return(indexes2)
}

Economic Indicators

Lastly, we will pull the macroeconomic indicator data from the csv files we have gathered from the World Bank Macroeconomic Indicators database.  The "gatherMacroData" function fits all of our macroeconomic indicators because the data was downloaded in a standardized format.  We will be making a function for each indicator so that we can import these functions in our analysis script. As always, we set NA for missing data.

%%R

gatherMacroData = function(filename) {
  macro = read.csv(filename)
  macro[1] = NULL
  macro[2] = NULL
  macro[2] = NULL
  names(macro)[2:20] = seq(1995, 2013)
  macro = t(macro)
  macro2 = macro[-1,]
  colnames(macro2) = macro[1,]
  macro3 = macro2[rowSums(is.na(macro2)) != ncol(macro2),]
  macro4 = apply(as.matrix(macro3), 1, as.numeric)
  macro5 = t(macro4)
  colnames(macro5) = tolower(macro[1,])
  macro6 = macro5[,colSums(is.na(macro5)) != nrow(macro5)]
  return(macro6)
}

getCO2Emissions = function() {
  return(gatherMacroData("./data/macros/co2emissions.csv"))
}
getExportsGrowth = function() {
  return(gatherMacroData("./data/macros/exportsGSGrowth.csv"))
}
getGDPGrowth = function() {
  return(gatherMacroData("./data/macros/gdpGrowth.csv"))
}
getGDSaving = function() {
  return(gatherMacroData("./data/macros/gdSaving.csv"))
}
getInfantMortality = function() {
  return(gatherMacroData("./data/macros/infantMortality.csv"))
}
getNetMigration = function() {
  return(gatherMacroData("./data/macros/netMigration.csv"))
}
getRealInterestRates = function() {
  return(gatherMacroData("./data/macros/realInterestRate.csv"))
}

Conclusion

Our data cleaning process loads the .csv files previously produced from data gathering into R Data Frames, with some cleaning for certain datasets along the way. We decided to store everything into R Data Frames because this is the easiest R object to carry out analysis moving forward. In summary, we know have a clean 2-D R Data Frames for each of our 4 datasets:
<br />
<br />
1. **TDI Data Frame**: rows are years 1995-2011; columns are countries in 3-letter standard country codes; values are TDI for the country in that year.
<br />
2. **Currencies Data Frame**: rows are years 1995-2011; columns are countries in 3-letter standard country codes; values are **normalized** exchange rate averages with 1 USD for the country's currency in that year.
<br />
3. **Stock Index Data Frame**: rows are years 1995-2011; columns are countries in 3-letter standard country codes; values are **normalized** stock index averages in that country for the year.
<br />
4. **Economic Indicators Data Frame**: rows are years 1995-2011; columns are countries in 3-letter standard country codes; values are the respective macro-economic indicator score/metric.
<br />
<br />
As we can see above, even though our datasets and sources are varied, we have cleaned them into R Data Frames having the same rows and columns for easy analysis in the next part.

References

* The Observatory of Economic Complexity. 
    * Accessed with http://atlas.media.mit.edu/

* The World Bank: World Development Indicators. 
    * Accessed with http://databank.worldbank.org/data/views/variableSelection/selectvariables.aspx?source=world-development-indicators

* Oanda Currency Converter. 
    * Accessed with http://www.oanda.com/currency/converter/
